round(number[, ndigits]) function in python 2 returns floating point number, which results in sda[position] throwing IndexError: : only integers, slices (:), ellipsis (...), numpy.newaxis (None) and integer or boolean arrays are valid indices

+++ Dcluster/rhodelta.py
@@ -10,7 +10,7 @@ def rhodelta(dist, xxdist, ND, N, percent = 2.0):
     print('Caculating rho and delta...')
     print('average percentage of neighbours (hard coded): %5.6f'%(percent) )
 
-    position = round(N*percent/100)
+    position = int(round(N*percent/100))
     sda=np.sort(xxdist)
     dc=sda[position]
     print('Computing Rho with gaussian kernel of radius: %12.6f\n'%(dc))
